+++ tests/frontend/org/voltdb/regressionsuites/RegressionSuite.java
@@ -29,7 +29,6 @@ import java.math.MathContext;
 import java.math.RoundingMode;
 import java.net.ConnectException;
 import java.nio.channels.SocketChannel;
-import java.rmi.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -80,19 +79,19 @@ import junit.framework.TestCase;
  *
  */
 public class RegressionSuite extends TestCase {
-    protected final static double GEOGRAPHY_EPSILON = 1.0e-13;
+    final static double GEOGRAPHY_EPSILON = 1.0e-13;
 
-    protected static int m_verboseDiagnosticRowCap = 40;
+    static int m_verboseDiagnosticRowCap = 40;
     protected VoltServerConfig m_config;
     protected String m_username = "default";
     protected String m_password = "password";
     private final ArrayList<Client> m_clients = new ArrayList<>();
     private final ArrayList<SocketChannel> m_clientChannels = new ArrayList<>();
-    protected final String m_methodName;
+    final String m_methodName;
     // If the current RegressionSuite instance is the last one in the current VoltServerConfig,
     // shutdown the cluster completely after finishing the test.
-    protected boolean m_completeShutdown;
-    protected boolean m_fatalFailure;
+    boolean m_completeShutdown;
+    boolean m_fatalFailure;
 
     /**
      * Trivial constructor that passes parameter on to superclass.
@@ -113,7 +112,6 @@ public class RegressionSuite extends TestCase {
      */
     @Override
     public void setUp() throws Exception {
-
         //New tests means a new server thread that hasn't done a restore
         m_config.setCallingMethodName(m_methodName);
         m_config.startUp(true);
@@ -140,11 +138,9 @@ public class RegressionSuite extends TestCase {
     public void tearDown() throws Exception {
         if (m_fatalFailure) {
             System.exit(0);
-        }
-        if (m_completeShutdown) {
+        } else if (m_completeShutdown) {
             m_config.shutDown();
-        }
-        else {
+        } else {
             Catalog currentCataog = getCurrentCatalog();
             if (currentCataog != null) {
                 CatalogDiffEngine diff = new CatalogDiffEngine(m_config.getInitialCatalog(), currentCataog);
@@ -172,8 +168,7 @@ public class RegressionSuite extends TestCase {
             for (String tableName : tableNames) {
                 try {
                     client.callProcedure("@AdHoc", "DELETE FROM " + tableName);
-                }
-                catch (ProcCallException pce) {
+                } catch (ProcCallException pce) {
                     if (! pce.getMessage().contains("Illegal to modify a materialized view.")) {
                         fail("Hit an exception when cleaning up tables between tests: " + pce.getMessage());
                     }
@@ -188,8 +183,7 @@ public class RegressionSuite extends TestCase {
             for (final SocketChannel sc : m_clientChannels) {
                 try {
                     ConnectionUtil.closeConnection(sc);
-                }
-                catch (final IOException e) {
+                } catch (final IOException e) {
                     e.printStackTrace();
                 }
             }
@@ -208,7 +202,7 @@ public class RegressionSuite extends TestCase {
     /**
      * @return The number of logical partitions in this configuration
      */
-    public int getLogicalPartitionCount() {
+    int getLogicalPartitionCount() {
         return m_config.getLogicalPartitionCount();
     }
 
@@ -219,7 +213,7 @@ public class RegressionSuite extends TestCase {
         return m_config.isValgrind();
     }
 
-    public boolean isLocalCluster() {
+    boolean isLocalCluster() {
         return m_config instanceof LocalCluster;
     }
 
@@ -230,7 +224,7 @@ public class RegressionSuite extends TestCase {
     /**
      * @return a reference to the associated VoltServerConfig
      */
-    public final VoltServerConfig getServerConfig() {
+    final VoltServerConfig getServerConfig() {
         return m_config;
     }
 
@@ -246,11 +240,11 @@ public class RegressionSuite extends TestCase {
         return getClient(1000 * 60 * 10, scheme); // 10 minute default
     }
 
-    public Client getClientToHostId(int hostId) throws IOException {
+    Client getClientToHostId(int hostId) throws IOException {
         return getClientToHostId(hostId, 1000 * 60 * 10); // 10 minute default
     }
 
-    public Client getClientToSubsetHosts(int[] hostIds) throws IOException {
+    Client getClientToSubsetHosts(int[] hostIds) throws IOException {
         return getClientToSubsetHosts(hostIds, 1000 * 60 * 10); // 10 minute default
     }
 
@@ -288,11 +282,10 @@ public class RegressionSuite extends TestCase {
 
     public Client getClient(long timeout, ClientAuthScheme scheme, boolean useAdmin) throws IOException {
         final Random r = new Random();
-        String listener = null;
+        String listener;
         if (useAdmin) {
             listener = m_config.getAdminAddress(r.nextInt(m_config.getListenerCount()));
-        }
-        else {
+        } else {
             listener = m_config.getListenerAddress(r.nextInt(m_config.getListenerCount()));
         }
         ClientConfig config = new ClientConfigForTest(m_username, m_password, scheme);
@@ -302,13 +295,10 @@ public class RegressionSuite extends TestCase {
         // Use the port generated by LocalCluster if applicable
         try {
             client.createConnection(listener);
-        }
-        // retry once
-        catch (ConnectException e) {
+        } catch (ConnectException e) { // retry once
             if (useAdmin) {
                 listener = m_config.getAdminAddress(r.nextInt(m_config.getListenerCount()));
-            }
-            else {
+            } else {
                 listener = m_config.getListenerAddress(r.nextInt(m_config.getListenerCount()));
             }
             client.createConnection(listener);
@@ -338,9 +328,7 @@ public class RegressionSuite extends TestCase {
         // Use the port generated by LocalCluster if applicable
         try {
             client.createConnection(listener);
-        }
-        // retry once
-        catch (ConnectException e) {
+        } catch (ConnectException e) { // retry once
             listener = listeners.get(r.nextInt(listeners.size()));
             client.createConnection(listener);
         }
@@ -355,7 +343,7 @@ public class RegressionSuite extends TestCase {
      * @return A VoltClient instance connected to the server driven by the
      * VoltServerConfig instance.
      */
-    public Client getClientToHostId(int hostId, long timeout) throws IOException {
+    Client getClientToHostId(int hostId, long timeout) throws IOException {
         final String listener = m_config.getListenerAddress(hostId);
         ClientConfig config = new ClientConfigForTest(m_username, m_password);
         config.setConnectionResponseTimeout(timeout);
@@ -363,16 +351,14 @@ public class RegressionSuite extends TestCase {
         final Client client = ClientFactory.createClient(config);
         try {
             client.createConnection(listener);
-        }
-        // retry once
-        catch (ConnectException e) {
+        } catch (ConnectException e) { // retry once
             client.createConnection(listener);
         }
         m_clients.add(client);
         return client;
     }
 
-    public Client getClientToSubsetHosts(int[] hostIds, long timeout) throws IOException {
+    Client getClientToSubsetHosts(int[] hostIds, long timeout) throws IOException {
         List<String> listeners = new ArrayList<>();
         for (int hostId : hostIds) {
             listeners.add(m_config.getListenerAddress(hostId));
@@ -384,9 +370,7 @@ public class RegressionSuite extends TestCase {
         for (String listener : listeners) {
             try {
                 client.createConnection(listener);
-            }
-            // retry once
-            catch (ConnectException e) {
+            } catch (ConnectException e) { // retry once
                 client.createConnection(listener);
             }
         }
@@ -405,9 +389,7 @@ public class RegressionSuite extends TestCase {
             // Use the port generated by LocalCluster if applicable
             try {
                 client.createConnection(listener);
-            }
-            // retry once
-            catch (ConnectException e) {
+            } catch (ConnectException e) { // retry once
                 listener = listeners.get(r.nextInt(listeners.size()));
                 client.createConnection(listener);
             }
@@ -419,7 +401,7 @@ public class RegressionSuite extends TestCase {
     /**
      * Release a client instance and any resources associated with it
      */
-    public void releaseClient(Client c) throws InterruptedException {
+    void releaseClient(Client c) throws InterruptedException {
         boolean removed = m_clients.remove(c);
         assert(removed);
         c.close();
@@ -433,11 +415,11 @@ public class RegressionSuite extends TestCase {
      * @return A SocketChannel that is already authenticated with the server
      */
 
-    public SocketChannel getClientChannel() throws IOException {
+    SocketChannel getClientChannel() throws IOException {
         return getClientChannel(false);
     }
 
-    public SocketChannel getClientChannel(final boolean noTearDown) throws IOException {
+    SocketChannel getClientChannel(final boolean noTearDown) throws IOException {
         final List<String> listeners = m_config.getListenerAddresses();
         final Random r = new Random();
         final String listener = listeners.get(r.nextInt(listeners.size()));
@@ -511,7 +493,7 @@ public class RegressionSuite extends TestCase {
         return isLocalCluster() ? ((LocalCluster)m_config).internalPort(hostId) : org.voltcore.common.Constants.DEFAULT_INTERNAL_PORT+hostId;
     }
 
-    static protected void validateDMLTupleCount(Client c, String sql, long modifiedTupleCount)
+    static void validateDMLTupleCount(Client c, String sql, long modifiedTupleCount)
             throws IOException, ProcCallException {
         VoltTable vt = c.callProcedure("@AdHoc", sql).getResults()[0];
         validateTableOfLongs(sql, vt, new long[][] {{modifiedTupleCount}});
@@ -539,7 +521,7 @@ public class RegressionSuite extends TestCase {
         validateTableOfScalarLongs(vt, expected);
     }
 
-    static protected void validateTableOfScalarDecimals(Client client, String sql, BigDecimal[] expected)
+    static void validateTableOfScalarDecimals(Client client, String sql, BigDecimal[] expected)
             throws IOException, ProcCallException {
         assertNotNull(expected);
         VoltTable vt = client.callProcedure("@AdHoc", sql).getResults()[0];
@@ -572,16 +554,12 @@ public class RegressionSuite extends TestCase {
         }
     }
 
-    private static void validateTableOfLongs(String messagePrefix,
-            VoltTable vt, long[][] expected) {
+    private static void validateTableOfLongs(String messagePrefix, VoltTable vt, long[][] expected) {
         assertNotNull(expected);
-        if (expected.length != vt.getRowCount()) {
-            if (vt.getRowCount() < m_verboseDiagnosticRowCap) {
-                System.out.println("Diagnostic dump of unexpected result for " + messagePrefix + " : " + vt);
-                System.out.println("VS. expected : ");
-                dumpExpectedLongs(expected);
-            }
-            //* enable and set breakpoint to debug multiple row count mismatches and continue */ return;
+        if (expected.length != vt.getRowCount() && vt.getRowCount() < m_verboseDiagnosticRowCap) {
+            System.out.println("Diagnostic dump of unexpected result for " + messagePrefix + " : " + vt);
+            System.out.println("VS. expected : ");
+            dumpExpectedLongs(expected);
         }
         assertEquals(messagePrefix + " returned wrong number of rows.  ",
                         expected.length, vt.getRowCount());
@@ -591,12 +569,10 @@ public class RegressionSuite extends TestCase {
         }
     }
 
-    protected void validateRowCount(Client client, String query, int expected)
-            throws NoConnectionsException, IOException, ProcCallException {
+    void validateRowCount(Client client, String query, int expected) throws IOException, ProcCallException {
         VoltTable result = client.callProcedure("@AdHoc", query).getResults()[0];
         int actual = result.getRowCount();
-        assertEquals("Wrong row count from query '" + query + "'",
-                expected, actual);
+        assertEquals("Wrong row count from query '" + query + "'", expected, actual);
     }
 
     public static void validateTableOfLongs(VoltTable vt, long[][] expected) {
@@ -613,25 +589,20 @@ public class RegressionSuite extends TestCase {
             // ENG-4295: hsql bug: HSQLBackend sometimes returns wrong column type.
             try {
                 actual = vt.getLong(i);
-            }
-            catch (IllegalArgumentException ex) {
+            } catch (IllegalArgumentException ex) {
                 try {
                     actual = (long) vt.getDouble(i);
-                }
-                catch (IllegalArgumentException newEx) {
+                } catch (IllegalArgumentException newEx) {
                     try {
                         actual = vt.getTimestampAsLong(i);
-                    }
-                    catch (IllegalArgumentException exTm) {
+                    } catch (IllegalArgumentException exTm) {
                         try {
                             actual = vt.getDecimalAsBigDecimal(i).longValueExact();
-                        }
-                        catch (IllegalArgumentException newerEx) {
+                        } catch (IllegalArgumentException newerEx) {
                             newerEx.printStackTrace();
                             fail(message);
                         }
-                    }
-                    catch (ArithmeticException newestEx) {
+                    } catch (ArithmeticException newestEx) {
                         newestEx.printStackTrace();
                         fail(message);
                     }
@@ -660,7 +631,7 @@ public class RegressionSuite extends TestCase {
      *
      * @param input
      */
-    static protected <T> T[][] shuffleArray(T [][] input) {
+    static<T> T[][] shuffleArray(T [][] input) {
         T[][] output = input.clone();
         Integer [] indices = new Integer[input.length];
         for (int idx = 0; idx < indices.length; idx += 1) {
@@ -674,7 +645,7 @@ public class RegressionSuite extends TestCase {
         return output;
     }
 
-    static protected void validateTableColumnOfScalarLong(VoltTable vt, int col, long[] expected) {
+    static void validateTableColumnOfScalarLong(VoltTable vt, int col, long[] expected) {
         assertNotNull(expected);
         assertEquals(expected.length, vt.getRowCount());
         int len = expected.length;
@@ -685,15 +656,14 @@ public class RegressionSuite extends TestCase {
             if (expected[i] == Long.MIN_VALUE) {
                 assertTrue(vt.wasNull());
                 assertEquals(null, actual);
-            }
-            else {
+            } else {
                 assertEquals(expected[i], actual);
             }
         }
     }
 
     static protected void validateTableColumnOfScalarVarchar(Client client, String sql, String[] expected)
-            throws NoConnectionsException, IOException, ProcCallException {
+            throws IOException, ProcCallException {
         VoltTable vt = client.callProcedure("@AdHoc", sql).getResults()[0];
         validateTableColumnOfScalarVarchar(vt, 0, expected);
     }
@@ -711,16 +681,15 @@ public class RegressionSuite extends TestCase {
             if (expected[i] == null) {
                 String actual = vt.getString(col);
                 assertTrue(vt.wasNull());
-                assertEquals(null, actual);
-            }
-            else {
+                assertNull(actual);
+            } else {
                 assertEquals(expected[i], vt.getString(col));
             }
         }
     }
 
 
-    static protected void validateTableColumnOfScalarVarbinary(Client client, String sql, String[] expected)
+    static void validateTableColumnOfScalarVarbinary(Client client, String sql, String[] expected)
             throws IOException, ProcCallException {
         VoltTable vt = client.callProcedure("@AdHoc", sql).getResults()[0];
         validateTableColumnOfScalarVarbinary(vt, 0, expected);
@@ -736,21 +705,20 @@ public class RegressionSuite extends TestCase {
 
               if (expected[i] == null) {
                   assertTrue(vt.wasNull());
-                  assertEquals(null, actual);
-              }
-              else {
+                  assertNull(actual);
+              } else {
                   assertEquals(expected[i], Encoder.hexEncode(actual));
               }
           }
     }
 
-    static protected void validateTableColumnOfScalarFloat(Client client, String sql, double[] expected)
+    static void validateTableColumnOfScalarFloat(Client client, String sql, double[] expected)
             throws IOException, ProcCallException {
         VoltTable vt = client.callProcedure("@AdHoc", sql).getResults()[0];
         validateTableColumnOfScalarFloat(vt, 0, expected);
     }
 
-    static protected void validateTableColumnOfScalarFloat(VoltTable vt, int col, double[] expected) {
+    static void validateTableColumnOfScalarFloat(VoltTable vt, int col, double[] expected) {
           assertNotNull(expected);
           assertEquals(expected.length, vt.getRowCount());
           int len = expected.length;
@@ -760,8 +728,7 @@ public class RegressionSuite extends TestCase {
 
               if (expected[i] <= VoltType.NULL_FLOAT) {
                   assertTrue(vt.wasNull());
-              }
-              else {
+              } else {
                   assertEquals(expected[i], actual, 0.00001);
               }
           }
@@ -774,29 +741,24 @@ public class RegressionSuite extends TestCase {
             BigDecimal actual = null;
             try {
                 actual = vt.getDecimalAsBigDecimal(i);
-            }
-            catch (IllegalArgumentException ex) {
+            } catch (IllegalArgumentException ex) {
                 ex.printStackTrace();
                 fail();
             }
             if (expected[i] != null) {
                 assertNotSame(null, actual);
                 assertEquals(expected[i], actual);
-            }
-            else {
-                if (isHSQL()) {
-                    // We don't actually use this with
-                    // HSQL.  So, just assert failure here.
-                    fail("HSQL is not used to test the Volt DECIMAL type.");
-                }
-                else {
-                    assertTrue(vt.wasNull());
-                }
+            } else if (isHSQL()) {
+                // We don't actually use this with
+                // HSQL.  So, just assert failure here.
+                fail("HSQL is not used to test the Volt DECIMAL type.");
+            } else {
+                assertTrue(vt.wasNull());
             }
         }
     }
 
-    protected void validateTableOfDecimal(VoltTable vt, BigDecimal[][] expected) {
+    private void validateTableOfDecimal(VoltTable vt, BigDecimal[][] expected) {
         assertNotNull(expected);
         assertEquals(expected.length, vt.getRowCount());
         int len = expected.length;
@@ -805,17 +767,17 @@ public class RegressionSuite extends TestCase {
         }
     }
 
-    protected static int m_defaultScale = 12;
-    protected static String m_roundingEnabledProperty = "BIGDECIMAL_ROUND";
-    protected static String m_roundingModeProperty = "BIGDECIMAL_ROUND_POLICY";
-    protected static String m_defaultRoundingEnablement = "true";
-    protected static String m_defaultRoundingMode = "HALF_UP";
+    static int m_defaultScale = 12;
+    static String m_roundingEnabledProperty = "BIGDECIMAL_ROUND";
+    static String m_roundingModeProperty = "BIGDECIMAL_ROUND_POLICY";
+    static String m_defaultRoundingEnablement = "true";
+    static String m_defaultRoundingMode = "HALF_UP";
 
-    protected static String getRoundingString(String label) {
+    static String getRoundingString(String label) {
         return String.format("%sRounding %senabled, mode is %s",
-                             label == null ? (label + ": ") : "",
-                             VoltDecimalHelper.isRoundingEnabled() ? "is " : "is *NOT* ",
-                             VoltDecimalHelper.getRoundingMode().toString());
+                label == null ? (label + ": ") : "",
+                VoltDecimalHelper.isRoundingEnabled() ? "is " : "is *NOT* ",
+                VoltDecimalHelper.getRoundingMode().toString());
     }
 
     /*
@@ -824,7 +786,7 @@ public class RegressionSuite extends TestCase {
      * using the given mode.  If roundingEnabled is false, no
      * rounding is done.
      */
-    protected static final BigDecimal roundDecimalValue(
+    static BigDecimal roundDecimalValue(
             String decimalValueString, boolean roundingEnabled, RoundingMode mode) {
         BigDecimal bd = new BigDecimal(decimalValueString);
         if (!roundingEnabled) {
@@ -847,19 +809,19 @@ public class RegressionSuite extends TestCase {
         return nbd;
     }
 
-    protected void validateTableOfDecimal(Client c, String sql, BigDecimal[][] expected) throws Exception {
+    void validateTableOfDecimal(Client c, String sql, BigDecimal[][] expected) throws Exception {
         assertNotNull(expected);
         VoltTable vt = c.callProcedure("@AdHoc", sql).getResults()[0];
         validateTableOfDecimal(vt, expected);
     }
 
-    static protected void validateTableColumnOfScalarDecimal(Client client, String sql, BigDecimal[] expected)
+    static void validateTableColumnOfScalarDecimal(Client client, String sql, BigDecimal[] expected)
             throws IOException, ProcCallException {
         VoltTable vt = client.callProcedure("@AdHoc", sql).getResults()[0];
         validateTableColumnOfScalarDecimal(vt, 0, expected);
     }
 
-    static protected void validateTableColumnOfScalarDecimal(VoltTable vt, int col, BigDecimal[] expected) {
+    static void validateTableColumnOfScalarDecimal(VoltTable vt, int col, BigDecimal[] expected) {
         assertNotNull(expected);
         assertEquals(expected.length, vt.getRowCount());
         int len = expected.length;
@@ -869,21 +831,20 @@ public class RegressionSuite extends TestCase {
 
             if (expected[i] == null) {
                 assertTrue(vt.wasNull());
-                assertEquals(null, actual);
-            }
-            else {
+                assertNull(actual);
+            } else {
                 BigDecimal rounded = expected[i].setScale(m_defaultScale, RoundingMode.valueOf(m_defaultRoundingMode));
                 assertEquals(rounded, actual);
             }
         }
   }
 
-    protected void assertTablesAreEqual(String prefix, VoltTable expectedRows, VoltTable actualRows) {
+    void assertTablesAreEqual(String prefix, VoltTable expectedRows, VoltTable actualRows) {
         assertTablesAreEqual(prefix, expectedRows, actualRows, null);
     }
 
     private static final long TOO_MUCH_INFO = 100;
-    protected void assertTablesAreEqual(String prefix, VoltTable expectedRows, VoltTable actualRows, Double epsilon) {
+    void assertTablesAreEqual(String prefix, VoltTable expectedRows, VoltTable actualRows, Double epsilon) {
         assertEquals(prefix + "column count mismatch.  Expected: " + expectedRows.getColumnCount() + " actual: " + actualRows.getColumnCount(),
                 expectedRows.getColumnCount(), actualRows.getColumnCount());
         if (expectedRows.getRowCount() != actualRows.getRowCount()) {
@@ -892,8 +853,7 @@ public class RegressionSuite extends TestCase {
             if (expRowCount + actRowCount < TOO_MUCH_INFO) {
                 System.out.println("Expected: " + expectedRows);
                 System.out.println("Actual:   " + actualRows);
-            }
-            else {
+            } else {
                 System.out.println("Expected: " + expRowCount + " rows");
                 System.out.println("Actual:   " + actRowCount + " rows");
             }
@@ -918,8 +878,7 @@ public class RegressionSuite extends TestCase {
                         continue;
                     }
                     fail(colPrefix + "expected null, got non null value: " + actualObj);
-                }
-                else {
+                } else {
                     assertFalse(colPrefix + "expected the value " + expectedObj +
                             ", got a null value.",
                             actualRows.wasNull());
@@ -947,8 +906,9 @@ public class RegressionSuite extends TestCase {
     }
 
     public static void assertEquals(String msg, GeographyPointValue expected, GeographyPointValue actual) {
-            assertApproximatelyEquals(msg, expected, actual, GEOGRAPHY_EPSILON);
+        assertApproximatelyEquals(msg, expected, actual, GEOGRAPHY_EPSILON);
     }
+
     /**
      * Assert that two points are approximately equal.  By this we mean the latitude and
      * longitude differ by at most epsilon.  If epsilon is zero or negative this means
@@ -962,8 +922,7 @@ public class RegressionSuite extends TestCase {
         if (epsilon > 0) {
             assertEquals(msg + " latitude: ", expected.getLatitude(), actual.getLatitude(), epsilon);
             assertEquals(msg + " longitude: ", expected.getLongitude(), actual.getLongitude(), epsilon);
-        }
-        else {
+        } else {
             assertEquals(msg + " latitude: ", expected.getLatitude(), actual.getLatitude());
             assertEquals(msg + " longitude: ", expected.getLongitude(), actual.getLongitude());
         }
@@ -996,7 +955,7 @@ public class RegressionSuite extends TestCase {
      * @param actual
      * @param epsilon
      */
-    public static void assertApproximatelyEquals(String msg, GeographyValue expected, GeographyValue actual, double epsilon) {
+    static void assertApproximatelyEquals(String msg, GeographyValue expected, GeographyValue actual, double epsilon) {
         if (expected == actual) {
             return;
         }
@@ -1040,10 +999,7 @@ public class RegressionSuite extends TestCase {
         assertEquals("Geographies not equal: ", expected, actual);
     }
 
-    private static void assertApproximateContentOfRow(int row,
-                                                      Object[] expectedRow,
-                                                      VoltTable actualRow,
-                                                      double epsilon) {
+    private static void assertApproximateContentOfRow(int row, Object[] expectedRow, VoltTable actualRow, double epsilon) {
         assertEquals("Actual row has wrong number of columns",
                 expectedRow.length, actualRow.getColumnCount());
         for (int i = 0; i < expectedRow.length; ++i) {
@@ -1061,30 +1017,23 @@ public class RegressionSuite extends TestCase {
                             + Arrays.toString((Byte[])actualValue);
                 }
                 assertTrue(fullMsg, actualRow.wasNull());
-            }
-            else if (expectedObj instanceof GeographyPointValue) {
+            } else if (expectedObj instanceof GeographyPointValue) {
                 assertApproximatelyEquals(msg, (GeographyPointValue) expectedObj, actualRow.getGeographyPointValue(i), epsilon);
-            }
-            else if (expectedObj instanceof GeographyValue) {
+            } else if (expectedObj instanceof GeographyValue) {
                 assertApproximatelyEquals(msg, (GeographyValue) expectedObj, actualRow.getGeographyValue(i), epsilon);
-            }
-            else if (expectedObj instanceof Long) {
-                long val = ((Long)expectedObj).longValue();
+            } else if (expectedObj instanceof Long) {
+                long val = (Long)expectedObj;
                 assertEquals(msg, val, actualRow.getLong(i));
-            }
-            else if (expectedObj instanceof Integer) {
+            } else if (expectedObj instanceof Integer) {
                 long val = ((Integer)expectedObj).longValue();
                 assertEquals(msg, val, actualRow.getLong(i));
-            }
-            else if (expectedObj instanceof Short) {
+            } else if (expectedObj instanceof Short) {
                 long val = ((Short)expectedObj).longValue();
                 assertEquals(msg, val, actualRow.getLong(i));
-            }
-            else if (expectedObj instanceof Byte) {
+            } else if (expectedObj instanceof Byte) {
                 long val = ((Byte)expectedObj).longValue();
                 assertEquals(msg, val, actualRow.getLong(i));
-            }
-            else if (expectedObj instanceof Double) {
+            } else if (expectedObj instanceof Double) {
                 Double expectedValue = (Double)expectedObj;
                 double actualValue = actualRow.getDouble(i);
                 // Either both are null or neither is null
@@ -1099,40 +1048,34 @@ public class RegressionSuite extends TestCase {
                                                          Math.abs(expectedValue - actualValue), epsilon);
                     assertTrue(fullMsg, Math.abs(expectedValue - actualValue) < epsilon);
                 }
-            }
-            else if (expectedObj instanceof BigDecimal) {
+            } else if (expectedObj instanceof BigDecimal) {
                 BigDecimal exp = (BigDecimal)expectedObj;
                 BigDecimal got = actualRow.getDecimalAsBigDecimal(i);
                 // Either both are null or neither is null
                 assertEquals(msg+"expected "+exp+" but got "+got+": checking for null DECIMAL: ",
                         exp == null, got == null);
                 assertEquals(msg, exp.doubleValue(), got.doubleValue(), epsilon);
-            }
-            else if (expectedObj instanceof byte[]) {
+            } else if (expectedObj instanceof byte[]) {
                 byte[] expectedVarbinary = (byte[]) expectedObj;
                 byte[] actualVarbinary = actualRow.getVarbinary(i);
                 assertEquals(msg+"length of VARBINARY: ", expectedVarbinary.length, actualVarbinary.length);
                 for (int k = 0; k < expectedVarbinary.length; k++) {
                     assertEquals(msg+"index "+k+" of VARBINARY value: ", expectedVarbinary[k], actualVarbinary[k]);
                 }
-            }
-            else if (expectedObj instanceof Byte[]) {
+            } else if (expectedObj instanceof Byte[]) {
                 Byte[] expectedVarbinary = (Byte[]) expectedObj;
                 Byte[] actualVarbinary = SerializationHelper.boxUpByteArray(actualRow.getVarbinary(i));
                 assertEquals(msg+"length of VARBINARY: ", expectedVarbinary.length, actualVarbinary.length);
                 for (int k = 0; k < expectedVarbinary.length; k++) {
                     assertEquals(msg+"index "+k+" of VARBINARY value: ", expectedVarbinary[k], actualVarbinary[k]);
                 }
-            }
-            else if (expectedObj instanceof String) {
+            } else if (expectedObj instanceof String) {
                 String val = (String)expectedObj;
                 assertEquals(msg, val, actualRow.getString(i));
-            }
-            else if (expectedObj instanceof TimestampType) {
+            } else if (expectedObj instanceof TimestampType) {
                 TimestampType val = (TimestampType)expectedObj;
                 assertEquals(msg, val, actualRow.getTimestampAsTimestamp(i));
-            }
-            else {
+            } else {
                 fail("Unexpected type in expected row: " + expectedObj.getClass().getSimpleName());
             }
         }
@@ -1156,9 +1099,7 @@ public class RegressionSuite extends TestCase {
      * @param actualTable
      * @param epsilon
      */
-    public static void assertApproximateContentOfTable(Object[][] expectedTable,
-                                                       VoltTable actualTable,
-                                                       double epsilon) {
+    static void assertApproximateContentOfTable(Object[][] expectedTable, VoltTable actualTable, double epsilon) {
         for (int i = 0; i < expectedTable.length; ++i) {
             assertTrue("Fewer rows than expected: "
                     + "expected: " + expectedTable.length + ", "
@@ -1187,7 +1128,7 @@ public class RegressionSuite extends TestCase {
         verifyProcFails(client, expectedPattern, "@AdHoc", stmt);
     }
 
-    static protected void verifyAdHocFails(Client client, String expectedPattern, Object... args) throws IOException {
+    static void verifyAdHocFails(Client client, String expectedPattern, Object... args) throws IOException {
         verifyProcFails(client, expectedPattern, "@AdHoc", args);
     }
 
@@ -1196,15 +1137,13 @@ public class RegressionSuite extends TestCase {
         String what;
         if (storedProc.compareTo("@AdHoc") == 0) {
             what = "the statement \"" + args[0] + "\"";
-        }
-        else {
+        } else {
             what = "the stored procedure \"" + storedProc + "\"";
         }
 
         try {
             client.callProcedure(storedProc, args);
-        }
-        catch (ProcCallException pce) {
+        } catch (ProcCallException pce) {
             String msg = pce.getMessage();
             String diagnostic = "Expected " + what + " to throw an exception matching the pattern \"" +
                     expectedPattern + "\", but instead it threw an exception containing \"" + msg + "\".";
@@ -1223,8 +1162,7 @@ public class RegressionSuite extends TestCase {
     // THE VOLTDB DOCS, RATHER THAN REUSING THE CODE THAT GENERATES THEM.
     // IN SOME MAGICAL FUTURE MAYBE THEY ALL CAN BE GENERATED FROM THE
     // SAME METADATA.
-    static protected void validateSchema(VoltTable result, VoltTable expected)
-    {
+    static protected void validateSchema(VoltTable result, VoltTable expected) {
         assertEquals(expected.getColumnCount(), result.getColumnCount());
         for (int i = 0; i < result.getColumnCount(); i++) {
             assertEquals("Failed name column: " + i, expected.getColumnName(i), result.getColumnName(i));
@@ -1236,7 +1174,8 @@ public class RegressionSuite extends TestCase {
         validStatisticsForTableLimitAndPercentage(client, tableName, limit, -1);
     }
 
-    static protected void validStatisticsForTableLimitAndPercentage(Client client, String tableName, long limit, long percentage) throws Exception {
+    static void validStatisticsForTableLimitAndPercentage(Client client, String tableName, long limit, long percentage)
+            throws Exception {
         long start = System.currentTimeMillis();
         while (true) {
             long lastLimit =-1, lastPercentage = -1;
@@ -1416,7 +1355,7 @@ public class RegressionSuite extends TestCase {
      * @param client
      * @throws Exception
      */
-    public void dropAllFunctions(Client client) throws Exception {
+    private void dropAllFunctions(Client client) throws Exception {
         dropAllTheThings(client, "function");
     }
 
@@ -1426,7 +1365,7 @@ public class RegressionSuite extends TestCase {
      * @param client
      * @throws Exception
      */
-    public void dropAllViews(Client client) throws Exception {
+    private void dropAllViews(Client client) throws Exception {
         dropAllTheThings(client, "view");
     }
 
@@ -1437,7 +1376,7 @@ public class RegressionSuite extends TestCase {
      * @param client
      * @throws Exception
      */
-    public void dropAllTables(Client client) throws Exception {
+    private void dropAllTables(Client client) throws Exception {
         dropAllTheThings(client, "table");
     }
 
@@ -1448,7 +1387,7 @@ public class RegressionSuite extends TestCase {
      * @param client
      * @throws Exception
      */
-    public void dropAllProcedures(Client client) throws Exception {
+    private void dropAllProcedures(Client client) throws Exception {
         dropAllTheThings(client, "procedure");
     }
 
@@ -1460,7 +1399,7 @@ public class RegressionSuite extends TestCase {
      * @param client
      * @throws Exception
      */
-    public void dropEverything(Client client) throws Exception {
+    void dropEverything(Client client) throws Exception {
         dropAllProcedures(client);
         dropAllViews(client);
         dropAllTables(client);
@@ -1478,7 +1417,7 @@ public class RegressionSuite extends TestCase {
      *        property-name-2, string-value-2, ...
      * @return the new Properties object
      **/
-    public static Properties buildProperties(String... alternatingKeysAndValues) {
+    protected static Properties buildProperties(String... alternatingKeysAndValues) {
         Properties properties = new Properties();
         int nStrings = alternatingKeysAndValues.length;
         // Each key should have a value, so the length should be even.
@@ -1504,15 +1443,14 @@ public class RegressionSuite extends TestCase {
                 int siteCount, int hostCount, int kFactor, boolean debug) {
 
             m_builder = new VoltProjectBuilder();
-            //Increase query tmeout as long literal queries taking long time.
+            //Increase query timeout as long literal queries taking long time.
             m_builder.setQueryTimeout(60000);
             try {
                 m_builder.addLiteralSchema(ddlText);
 
                 // add more partitioned and replicated tables, PARTED[1-3] and REPED[1-2]
                 AdHocQueryTester.setUpSchema(m_builder, pathToCatalog, pathToDeployment);
-            }
-            catch (Exception e) {
+            } catch (Exception e) {
                 e.printStackTrace();
                 fail("Failed to set up schema");
             }
@@ -1527,8 +1465,7 @@ public class RegressionSuite extends TestCase {
 
             try {
                 MiscUtils.copyFile(m_builder.getPathToDeployment(), pathToDeployment);
-            }
-            catch (Exception e) {
+            } catch (Exception e) {
                 fail(String.format("Failed to copy \"%s\" to \"%s\"", m_builder.getPathToDeployment(), pathToDeployment));
             }
         }
@@ -1540,9 +1477,6 @@ public class RegressionSuite extends TestCase {
                 // do the test
                 m_client = ClientFactory.createClient();
                 m_client.createConnection("localhost", m_cluster.port(0));
-            } catch (UnknownHostException e) {
-                e.printStackTrace();
-                fail(String.format("Failed to connect to localhost:%d", m_cluster.port(0)));
             } catch (IOException e) {
                 e.printStackTrace();
                 fail(String.format("Failed to connect to localhost:%d", m_cluster.port(0)));
